fix: resolve pre-existing compilation errors in PaymentServiceImpl and PaymentControllerTest
- PaymentServiceImpl: change feePercentage type from BigDecimal to double (getFeePercentage returns double)
- PaymentControllerTest: fix MockBean import path for Spring Boot 2.7.x (mockito subpackage)

+++ src/main/java/com/bigfake/payments/service/impl/PaymentServiceImpl.java
@@ -162,7 +162,7 @@ public class PaymentServiceImpl implements PaymentService {
         }
 
         // Calculate fees
-        BigDecimal feePercentage = request.getPaymentType().getFeePercentage();
+        double feePercentage = request.getPaymentType().getFeePercentage();
         BigDecimal feeAmount = request.getAmount()
                 .multiply(BigDecimal.valueOf(feePercentage))
                 .setScale(4, RoundingMode.HALF_UP);
